Extract color-mode/sRGB handling into engine/src/colorMode.ts
Moves pointColorsNeedSrgbDecode, setupPointSrgbDecode (in-shader sRGB
decode injection), buildOriginalColorArray, applyColorModeToGeometry,
shouldUseVertexColors, getColorName, and getColorOptions out of
PointCloudVisualizer. Only two functions need visualizer state
(convertSrgbToLinear, colorProcessor); the rest were already pure or
now call directly into the intensity/commentSettings modules instead
of going through `this`.

Also moved isDepthDerivedFile, isPngDerivedFile, and isRgbDerivedFile
into depth/commentSettings.ts alongside the other PLY-comment parsers
they belong with - noticed while tracing pointColorsNeedSrgbDecode's
one dependency.

main.ts: 12,330 -> 12,158 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
